+++ app/Services/CommissionService.php
@@ -10,13 +10,13 @@ final class CommissionService
     private int $user_id;
     private float $amount;
 
-    public function __construct(int $user_id, int $amount)
+    public function __construct(int $user_id, float $amount)
     {
         $this->user_id = $user_id;
         $this->amount = $amount;
     }
 
-    public function getFeeValue(): int
+    public function getFeeValue(): float
     {
         $feePercentage = $this->getCurrentFeePercentage($this->user_id);
         return $this->calculateFeeValue($this->amount, $feePercentage);
@@ -40,7 +40,7 @@ final class CommissionService
             : config('transaction.high_fee_percentage');
     }
 
-    private function calculateFeeValue(int $amount, int $fee_in_percentage): float
+    private function calculateFeeValue(float $amount, int $fee_in_percentage): float
     {
         return $amount * $fee_in_percentage / 100;
     }
